Move rename to legacy
